+++ mirinae/src/main/java/cs.xml
@@ -82,13 +82,7 @@
 		select count(*) from project
 	</select>
 	
-	<!-- 프로젝트 목록 최신순 -->
-	<select id="list3" resultType="project">
-		select * from
-		(select rowNum rn , a.* from
-		(select * from project order by p_no desc) a)
-		where rn between #{startRow} and #{endRow}
-	</select>
+	
 	
 	<!-- 관리자페이지 - 총 후원 수 -->
 	<select id="getTotalS" resultType="integer">
@@ -96,10 +90,5 @@
 	</select>
 	
 	<!-- 후원 목록 최신순 -->
-	<select id="list4" resultType="support">
-		select * from
-		(select rowNum rn , a.* from
-		(select * from support order by sup_no desc) a)
-		where rn between #{startRow} and #{endRow}
-	</select>
+	
 </mapper>
