+++ mirinae/src/main/webapp/WEB-INF/command_kwon.properties
@@ -13,3 +13,4 @@ views/admin/support_list.kwon=service.admin.SupportList
 views/admin/sup_view.kwon=service.admin.SupportView
 views/admin/sup_delete.kwon=service.admin.SupportDelete
 views/main/homeSearch.kwon=service.main.HomeSearch
+views/admin/p_del.kwon=service.admin.ProjectDel
