+++ mirinae/src/main/webapp/WEB-INF/views/admin/p_del.jsp
@@ -1,0 +1,24 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+    pageEncoding="UTF-8"%>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>Insert title here</title>
+</head>
+<body>
+<c:if test="${result > 0 }">
+	<script type="text/javascript">
+		alert("프로젝트 삭제가 완료되었습니다.");
+		location.href="project_list.kwon";
+	</script>
+</c:if>
+<c:if test="${result == 0 }">
+	<script type="text/javascript">
+		alert("프로젝트 삭제에 실패하셨습니다.");
+		history.go(-1);
+	</script>
+</c:if>
+</body>
+</html>
